+++ modules/utils/src/main/java/org/springside/modules/utils/collection/primitive/LongObjectHashMap.java
@@ -26,7 +26,9 @@ import java.util.Set;
 import org.springside.modules.utils.base.Maths;
 
 /**
- * 移植Netty 4.1.6的Key为原子类型的集合类, 在数据结构上与HashMap不一样，空间占用与性能俱比原来更优 
+ * 移植Netty 4.1.6的Key为原子类型的集合类, 在数据结构上与HashMap不一样，空间占用与读写性能俱比原来更优.
+ * 
+ * 原子类型集合类有多个实现，选择Netty是因为有在实战中使用.
  * 
  * A hash map implementation of {@link LongObjectMap} that uses open addressing for keys.
  * To minimize the memory footprint, this class uses open addressing rather than chaining.
@@ -188,7 +190,7 @@ public class LongObjectHashMap<V> implements LongObjectMap<V> {
 
     @Override
     public void clear() {
-        Arrays.fill(keys, (long) 0);
+        Arrays.fill(keys, 0);
         Arrays.fill(values, null);
         size = 0;
     }
@@ -328,7 +330,7 @@ public class LongObjectHashMap<V> implements LongObjectMap<V> {
     }
 
     private long objectToKey(Object key) {
-        return (long) ((Long) key).longValue();
+        return ((Long) key).longValue();
     }
 
     /**
